requirements.txt
matplotlib==2.1.2
scipy==1.0.0
numpy==1.14.0
Flask==1.0.2
torch==0.4.1
Pillow==5.2.0
torchvision==0.2.1
